Remove gtm e2e package and use e2e-analytics
