+++ e2e/packages/analytics/package.json
@@ -1,10 +1,11 @@
 {
   "name": "e2e-analytics",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "description": "Package to do e2e testing of Frontity's analytics library",
   "dependencies": {
     "@frontity/analytics": "^1.0.0-beta.0",
+    "@frontity/source": "^1.2.0",
     "frontity": "^1.4.3"
   }
 }
