+++ e2e/packages/analytics/src/index.tsx
@@ -1,73 +1,83 @@
 import React from "react";
-import {
-  Root as Analytics,
-  sendPageview,
-  sendEvent
-} from "@frontity/analytics";
-import { Head, connect } from "frontity";
+import { Head, connect, URL } from "frontity";
 import Package from "../types";
 
-const Path1 = () => {
-  const title = "";
-  return (
-    <>
-      <Head>
-        <title>{title}</title>
-      </Head>
-      <div>Lorem Ipsum</div>
-    </>
-  );
-};
+const Homepage = () => (
+  <>
+    <Head>
+      <title>Homepage Title</title>
+    </Head>
+    <h1>Homepage</h1>
+    <p>This is the homepage.</p>
+  </>
+);
 
-const Path2 = () => {
-  const title = "";
-  return (
-    <>
-      <Head>
-        <title>{title}</title>
-      </Head>
-      <div>Lorem Ipsum</div>
-    </>
-  );
-};
+const SomePost = () => (
+  <>
+    <Head>
+      <title>Some post Title</title>
+    </Head>
+    <h1>Some post</h1>
+    <p>This is some post</p>
+  </>
+);
 
-const Path3 = () => {
-  const title = "";
-  return (
-    <>
-      <Head>
-        <title>{title}</title>
-      </Head>
-      <div>Lorem Ipsum</div>
-    </>
-  );
-};
-
-const Theme = connect(({ state }) => {
+const Theme = connect(({ state, actions }) => {
+  // Get only the pathname (link has query).
   const { pathname } = new URL(state.router.link, "http://localhost:3001");
-  if (pathname === "/path-1") return <Path1 />;
-  if (pathname === "/path-2") return <Path2 />;
-  if (pathname === "/path-3") return <Path3 />;
+
+  const changeLink = () => actions.router.set("/some-post/");
+  const sendEvent = () => actions.analytics.sendEvent({ some: "content" });
+
+  return (
+    <>
+      {/* Render homepage or post */}
+      {pathname === "/" && <Homepage />}
+      {pathname === "/some-post/" && <SomePost />}
+      {/* Buttons */}
+      <button id="change-link" onClick={changeLink}>
+        Some Post
+      </button>
+      <button id="send-event" onClick={sendEvent}>
+        Send event
+      </button>
+    </>
+  );
 });
 
 const testPackage: Package = {
-  name: "analytics",
-  state: {
-    analytics: {
-      namespaces: {}
-    }
-  },
-  actions: {
-    analytics: {
-      // These actions could be mocked here.
-      sendPageview,
-      sendEvent
-    }
-  },
+  name: "e2e-analytics",
   roots: {
     theme: Theme,
-    analytics: Analytics
-  }
+  },
+  state: {
+    source: {
+      data: {},
+      get: ({ state }) => (link) => state.source.data[link],
+    },
+  },
+  actions: {
+    source: {
+      fetch: ({ state }) => async (link, options) => {
+        const { data } = state.source;
+        if (!data[link]) {
+          data[link] = {
+            link,
+            isFetching: true,
+            isReady: false,
+          };
+
+          await Promise.resolve();
+
+          data[link] = {
+            ...data[link],
+            isFetching: false,
+            isReady: true,
+          };
+        }
+      },
+    },
+  },
 };
 
 export default testPackage;
